feta: get user by id

+++ src/main/java/com/github/erneo97/medical_clinic/service/PatientService.java
@@ -18,7 +18,7 @@ public class PatientService {
         return inMemoryPatientRepository.findAll();
     }
 
-    public Optional<Patient> findById(int id) {
+    public Optional<Patient> findById(Long id) {
         return inMemoryPatientRepository.findById(id);
     }
